Make tests pass
