+++ clouds/tests/testSubject.py
@@ -7,8 +7,7 @@ import numpy as np
 
 from clouds.obj.subject import Subject
 from clouds.obj.classifier import Classifier
-from clouds.tests.testClassifier import testTrainClassifier
-
+from clouds.tests import util
 TESTDATA = './data'
 
 class testSubject(unittest.TestCase):
@@ -20,7 +19,7 @@ class testSubject(unittest.TestCase):
         cls.storedClassifier = os.path.join(cls.workspace, 'sc')
         shutil.copytree(os.path.join(TESTDATA, 'xorClassifier'), cls.storedClassifier)
 
-        cls.xors = testTrainClassifier.createXors(cls.workspace)
+        cls.xors = util.createXors(cls.workspace)
 
     @classmethod
     def tearDownClass(cls):
