+++ clouds/tests/util/__init__.py
@@ -1,0 +1,1 @@
+from .utils import createXors
